Update Activity log
This is an activity log. Here you can have 3 clients, and they can enter when they are doing as an activity, eating and exercising. The logged data will be recorded and stored. Later on, it can be tracked and retrieved. This system can be used to calculate daily intake and monthly intake as well by logging in whenever needed. Calorie burn and gain can be tracked, which can be used in a health facilitation service. When the diet consumed by the user and his exercise patterns are entered, the details of his activity are being saved in this activity log. This code is executed by Rishib Chakraborty and written in python programming language.